migrate duration class into span
